+++ decitala/cli.py
@@ -12,6 +12,9 @@ import json
 from decitala import __version__
 from .search import path_finder
 from .fragment import FragmentEncoder
+from .utils import get_logger
+
+logger = get_logger(name=__file__)
 
 @click.group()
 @click.version_option(__version__, "--version", "-v", message="%(version)s")
@@ -23,13 +26,17 @@ def decitala():
 @click.option("--filepath", default="", help="Path to filepath parsed for analysis.")
 @click.option("--part_num", default=0, help="Part number.")
 @click.option("--frag_type", default="greek_foot")
-def pathfinder(filepath, part_num, frag_type):
+@click.option("--verbose", default=True)
+def pathfinder(filepath, part_num, frag_type, verbose):
 	best_path = path_finder(
 		filepath=filepath,
 		part_num=part_num,
-		frag_type=frag_type
+		frag_type=frag_type,
+		verbose=verbose
 	)
-	filename = filepath.split("/")[-1][:-4] + "_part_num={0}_frag_type={1}.txt".format(part_num, frag_type) # noqa
+	json_dumped_res = json.dumps(obj=best_path, cls=FragmentEncoder, indent=4)
+	logger.info(json_dumped_res)
+	filename = filepath.split("/")[-1][:-4] + "_part_num={0}_frag_type={1}.json".format(part_num, frag_type) # noqa
 	with open(filename, "w") as output:
 		json.dump(obj=best_path, fp=output, cls=FragmentEncoder, indent=4)